LG-4253 Generate errors from acuant through error_generator to simplify the code and improve the user experience when there are doc auth errors.
