+++ lib/identity_doc_auth.rb
@@ -2,9 +2,11 @@ require "identity_doc_auth/acuant/acuant_client"
 require "identity_doc_auth/lexis_nexis/lexis_nexis_client"
 require "identity_doc_auth/mock/doc_auth_mock_client"
 require "identity_doc_auth/mock/result_response_builder"
+require "identity_doc_auth/error_generator"
+require "identity_doc_auth/errors"
+require "identity_doc_auth/request_error"
 require "identity_doc_auth/response"
 require "identity_doc_auth/version"
-require "identity_doc_auth/request_error"
 
 module IdentityDocAuth
 end
