+++ spec/fixtures/acuant_responses/get_results_response_failure.json
@@ -212,7 +212,7 @@
     "IsActive": true,
     "IsDevelopment": true,
     "IsTrial": false,
-    "Name": "Dev GSA iProov",
+    "Name": "Dev Test iProov",
     "StorePII": false
   }
 }
